+++ libs/oidc/src/utils/index.ts
@@ -2,4 +2,3 @@ export * from './merge-defaults.util';
 export * from './session.serializer';
 export * from './setup-session';
 export * from './oidc-helpers.util';
-export * from './setup-token-guard';