+++ libs/oidc/CHANGELOG.md
@@ -2,7 +2,38 @@
 
 **Table of Contents**
 
-<!-- TOC -->autoauto- [Changelog](#changelog)auto    - [0.7.0 (2020-06-)](#070-2020-06-)auto    - [0.6.1 (2020-06-03)](#061-2020-06-03)auto        - [Bugfixes](#bugfixes)auto    - [0.6.0 (2020-05-27)](#060-2020-05-27)auto        - [Features](#features)auto        - [Doc fixes](#doc-fixes)auto    - [0.5.0 (2020-05-19)](#050-2020-05-19)auto        - [Features](#features-1)auto        - [Bug fixes](#bug-fixes)auto        - [Documentation](#documentation)auto    - [0.4.0 (2020-05-06)](#040-2020-05-06)auto        - [Features](#features-2)auto    - [0.3.1 (2020-05-06)](#031-2020-05-06)auto        - [Bug fixes](#bug-fixes-1)auto    - [0.3.0 (2020-04-30)](#030-2020-04-30)auto        - [Features](#features-3)auto    - [0.2.1 (2020-04-30)](#021-2020-04-30)auto        - [Bug fixes](#bug-fixes-2)auto    - [0.2.0 (2020-04-23)](#020-2020-04-23)auto        - [Features](#features-4)auto            - [Before](#before)auto            - [Now](#now)auto    - [0.1.2 (2020-04-23)](#012-2020-04-23)auto        - [Bug fixes](#bug-fixes-3)autoauto<!-- /TOC -->
+<!-- TOC -->
+
+- [Changelog](#changelog)
+  - [0.7.0 (2020-06-05)](#070-2020-06-05)
+    - [Features](#features)
+      - [How to set a global guard BEFORE](#how-to-set-a-global-guard-before)
+      - [How to set a global guard NOW](#how-to-set-a-global-guard-now)
+  - [0.6.1 (2020-06-03)](#061-2020-06-03)
+    - [Bugfixes](#bugfixes)
+  - [0.6.0 (2020-05-27)](#060-2020-05-27)
+    - [Features](#features-1)
+    - [Doc fixes](#doc-fixes)
+  - [0.5.0 (2020-05-19)](#050-2020-05-19)
+    - [Features](#features-2)
+    - [Bug fixes](#bug-fixes)
+    - [Documentation](#documentation)
+  - [0.4.0 (2020-05-06)](#040-2020-05-06)
+    - [Features](#features-3)
+  - [0.3.1 (2020-05-06)](#031-2020-05-06)
+    - [Bug fixes](#bug-fixes-1)
+  - [0.3.0 (2020-04-30)](#030-2020-04-30)
+    - [Features](#features-4)
+  - [0.2.1 (2020-04-30)](#021-2020-04-30)
+    - [Bug fixes](#bug-fixes-2)
+  - [0.2.0 (2020-04-23)](#020-2020-04-23)
+    - [Features](#features-5)
+      - [Before](#before)
+      - [Now](#now)
+  - [0.1.2 (2020-04-23)](#012-2020-04-23)
+    - [Bug fixes](#bug-fixes-3)
+
+<!-- /TOC -->
 
 ## 0.7.0 (2020-06-05)